move PIV NACI indicator to its own module in prep for using new cms crate from formats repo instead of local definitions
