+++ certval/src/asn1.rs
@@ -1,5 +1,7 @@
 //! Sources of ASN.1 encoders and decoders not included in a RustCrypto formats repo
 
 pub mod cryptographic_message_syntax2004;
+pub mod piv_naci_indicator;
 
 pub use crate::asn1::cryptographic_message_syntax2004::*;
+pub use crate::asn1::piv_naci_indicator::*;
