+++ certval/src/validator/pdv_certificate.rs
@@ -7,7 +7,7 @@ use alloc::{
     vec::Vec,
 };
 
-use crate::asn1::cryptographic_message_syntax2004::PivNaciIndicator;
+use crate::asn1::piv_naci_indicator::PivNaciIndicator;
 use der::{
     asn1::{BitStringRef, ObjectIdentifier},
     Decode,
@@ -16,7 +16,7 @@ use spki::AlgorithmIdentifierOwned;
 use x509_cert::ext::{pkix::crl::CrlDistributionPoints, pkix::*};
 use x509_cert::*;
 
-use crate::asn1::cryptographic_message_syntax2004::PIV_NACI_INDICATOR;
+use crate::asn1::piv_naci_indicator::PIV_NACI_INDICATOR;
 use const_oid::db::rfc5912::{
     ID_CE_AUTHORITY_KEY_IDENTIFIER, ID_CE_BASIC_CONSTRAINTS, ID_CE_CERTIFICATE_POLICIES,
     ID_CE_CRL_DISTRIBUTION_POINTS, ID_CE_EXT_KEY_USAGE, ID_CE_ISSUER_ALT_NAME, ID_CE_KEY_USAGE,
